requirements.txt
numpy
opencv-python
pynput
screeninfo
mediapipe
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm
Pillow
playwright